Don't complete room transitions (in or out) while the game is paused
Previously, a CT_PAUSEGAME would stop all subsequent
actions from happening until the user hit 'Continue'.
If the pause happened on a roomleave, before any
CANCELMOVE actions, the room transition would succeed.
When the user hit 'continue' and the CANCELMOVE happened,
it wouldn't do anything.

Now, if the game is paused, the remaining room transition
steps are queued to ensure anything that was meant to happen
during the transition (a CANCELMOVE on exit, or showing
an image and pausing on enter) happens correctly.