Revert to original RAGS behavior where room image is shown before room events run
This is needed for Wife Trainer, when the intro room has a 'clear'
image, but the actual image to be set is in a 'On Enter' event.

You need the 'clear' image to be set first, then the image from
the 'On Enter', then nothing else until the user takes an action.