Make sure CT_MOVEPLAYER works even when the room is a name+sdesc combo
Though it usually produces a UniqueID, sometimes
when you manually edit a field in the Rags designer
(instead of selecting from a dropdown) the value
for a room (in e.x. CT_MOVEPLAYER) will be the
room name or `%{name}-%{sdesc}`. So we need to check for that.

Fixes error when doing supermarket work in The Sinner