Fix issue where room enter callbacks would not fully fire before next command
For example, if a command list had a 'CT_MOVEPLAYER' followed by a
'CT_PAUSEGAME', the game would get paused immediately after the
'On Enter First Time' hooks for that room, but before the
more important 'On Enter' hooks.

Refactor MasterCommandList into a CommandList object which
allows different 'stacks' of commands to be run.