CT_AdditionalDataCheck verifies the AdditionalData applies to the current condition
If multiple timers execute in a row that contain conditions that
want similar AdditionalData, it was possible for data chosen
in one action to be evaluated against the action that didn't
trigger it.

Now, all the commands/conditions are wrapped in a little object that
includes what their parent action was, so CT_AdditionalDataCheck
can decide to only return true if the AdditionalData collected
was actually for it.

Fixes opening perk selection in Superpowered