Variable comparisons against array ranges should always fail
RAGS can produce comparisons with operands like 'Array(2)(0..15)',
but it seems like in practice these always fail, even if the
sub-array is something like ['Hello', 'World'] and the
operation is asking if it contains 'Hello'. There might be
some circumstance where it passes, but it's hard to find!

Fixes issue with finishing a turn in No Haven