Fix issues with 'On Turn 1' timers not running when they should be
'Me Against the World' frequently use timers where the intended action
happens in the 'On Turn 1' command. it also often starts the timer like
this:
* Set Timer To Active [timername]
* Execute Timer

In the windows RAGS client, the 'execute' doesn't have any effect,
because as of the time it is run the timer is on turn 0.

The 'set timer to active' should make the 'On Turn 1' thing happen
next, but that's reliant on regalia successfully triggering the timer
event loop.

Prior to this version, if you triggered a character action that had a
pause, it would not trigger the timer event loop. Now, it should?